requirements.txt
Django>=3.1.13
psycopg2==2.8.6
psycopg2-binary==2.9.1
python-dotenv==0.15.0
djangorestframework==3.12.2
dj-database-url==0.5.0
django-on-heroku==1.1.2
django-heroku==0.3.1
gunicorn==20.0.4
djangorestframework-simplejwt==5.0.0
django-cors-headers==3.5.0
requests==2.25.1
Pillow==8.3.2